+++ packages/cli/src/index.ts
@@ -1,3 +1,5 @@
 export * from './commands/index.js'
 export * from './config.js'
+export * from './generation-handler.js'
 export * from './module-loader.js'
+export * from './utils.js'