+++ packages/generator-base/src/index.ts
@@ -1,0 +1,2 @@
+export * from './generator-type.js'
+export * from './generator.js'