+++ packages/generator-html-doc/src/html-doc-generator.ts
@@ -1,14 +1,13 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { Logger } from './logger.js'
-import { Generator } from './generator.js'
-import { DANGER_HTML_DOC_GENERATOR_NOT_IMPLEMENTED } from './messages.js'
-import type { GenerateConfig, GirModulesGrouped, InheritanceTable } from './types/index.js'
-import type { GirModule } from './gir-module.js'
+import { Logger, DANGER_HTML_DOC_GENERATOR_NOT_IMPLEMENTED } from '@ts-for-gir/lib'
+import { Generator } from '@ts-for-gir/generator-base'
+
+import type { GenerateConfig, GirModulesGrouped, InheritanceTable, GirModule } from '@ts-for-gir/lib'
 
 /**
  * A template that can be used to implement an HTML Documentation Generator
  */
-export default class HtmlDocGenerator implements Generator {
+export class HtmlDocGenerator implements Generator {
     protected log: Logger
     constructor(protected readonly config: GenerateConfig) {
         this.log = new Logger(config.environment, config.verbose, HtmlDocGenerator.name)